+++ profile_test.go
@@ -5,6 +5,8 @@ import (
 
 	. "github.com/onsi/gomega"
 	"golang.org/x/oauth2"
+
+	profileapi "github.com/dmalch/go-geni/profile"
 )
 
 func ptr[T any](s T) *T {
@@ -14,9 +16,9 @@ func ptr[T any](s T) *T {
 func TestCreateProfile1(t *testing.T) {
 	t.Skip()
 	RegisterTestingT(t)
-	profileRequest := ProfileRequest{
+	profileRequest := profileapi.Request{
 		Gender: ptr("male"),
-		Names: map[string]NameElement{
+		Names: map[string]profileapi.NameElement{
 			"en-US": {
 				FirstName: ptr("1TestFirstName"),
 				LastName:  ptr("1TestLastName"),
@@ -27,21 +29,21 @@ func TestCreateProfile1(t *testing.T) {
 				MiddleName: ptr("Н"),
 			},
 		},
-		Birth: &EventElement{
-			Date: &DateElement{
+		Birth: &profileapi.EventElement{
+			Date: &profileapi.DateElement{
 				Day:   ptr[int32](19),
 				Month: ptr[int32](8),
 				Year:  ptr[int32](1922),
 			},
-			Location: &LocationElement{
+			Location: &profileapi.LocationElement{
 				Country:   ptr("РСФСР"),
 				County:    ptr("Спасский уезд, Кирилловская волость"),
 				PlaceName: ptr("село Кирилово"),
 				State:     ptr("Тамбовская губерния"),
 			},
 		},
-		Death: &EventElement{
-			Date: &DateElement{
+		Death: &profileapi.EventElement{
+			Date: &profileapi.DateElement{
 				Day:   ptr[int32](25),
 				Month: ptr[int32](9),
 				Year:  ptr[int32](1993),
@@ -59,22 +61,22 @@ func TestCreateProfile1(t *testing.T) {
 	Expect(profile.FirstName).To(BeEquivalentTo("1TestFirstName"))
 	Expect(profile.LastName).To(BeEquivalentTo("1TestLastName"))
 	Expect(profile.Gender).To(BeEquivalentTo("male"))
-	Expect(profile.Names).To(HaveKeyWithValue("en-US", NameElement{
+	Expect(profile.Names).To(HaveKeyWithValue("en-US", profileapi.NameElement{
 		FirstName: ptr("1TestFirstName"),
 		LastName:  ptr("1TestLastName"),
 	}))
-	Expect(profile.Names).To(HaveKeyWithValue("ru", NameElement{
+	Expect(profile.Names).To(HaveKeyWithValue("ru", profileapi.NameElement{
 		FirstName:  ptr("Ф"),
 		LastName:   ptr("М"),
 		MiddleName: ptr("Н"),
 	}))
-	Expect(profile.Birth).To(Equal(&EventElement{
-		Date: &DateElement{
+	Expect(profile.Birth).To(Equal(&profileapi.EventElement{
+		Date: &profileapi.DateElement{
 			Day:   ptr[int32](19),
 			Month: ptr[int32](8),
 			Year:  ptr[int32](1922),
 		},
-		Location: &LocationElement{
+		Location: &profileapi.LocationElement{
 			Country:   ptr("РСФСР"),
 			County:    ptr("Спасский уезд, Кирилловская волость"),
 			PlaceName: ptr("село Кирилово"),
@@ -82,8 +84,8 @@ func TestCreateProfile1(t *testing.T) {
 		},
 		Name: "Birth of 1TestFirstName 1TestLastName",
 	}))
-	Expect(profile.Death).To(Equal(&EventElement{
-		Date: &DateElement{
+	Expect(profile.Death).To(Equal(&profileapi.EventElement{
+		Date: &profileapi.DateElement{
 			Day:   ptr[int32](25),
 			Month: ptr[int32](9),
 			Year:  ptr[int32](1993),
@@ -109,11 +111,11 @@ func TestGetProfile1(t *testing.T) {
 	Expect(profile.FirstName).To(BeEquivalentTo("D"))
 	Expect(profile.LastName).To(BeEquivalentTo("M"))
 	Expect(profile.Gender).To(BeEquivalentTo("male"))
-	Expect(profile.Names).To(HaveKeyWithValue("en-US", NameElement{
+	Expect(profile.Names).To(HaveKeyWithValue("en-US", profileapi.NameElement{
 		FirstName: ptr("D"),
 		LastName:  ptr("M"),
 	}))
-	Expect(profile.Names).To(HaveKeyWithValue("ru", NameElement{
+	Expect(profile.Names).To(HaveKeyWithValue("ru", profileapi.NameElement{
 		FirstName:  ptr("Д"),
 		LastName:   ptr("М"),
 		MiddleName: ptr("В"),
@@ -138,22 +140,22 @@ func TestGetProfile2(t *testing.T) {
 	Expect(profile.FirstName).To(BeEquivalentTo("F"))
 	Expect(profile.LastName).To(BeEquivalentTo("M"))
 	Expect(profile.Gender).To(BeEquivalentTo("male"))
-	Expect(profile.Names).To(HaveKeyWithValue("en-US", NameElement{
+	Expect(profile.Names).To(HaveKeyWithValue("en-US", profileapi.NameElement{
 		FirstName: ptr("F"),
 		LastName:  ptr("M"),
 	}))
-	Expect(profile.Names).To(HaveKeyWithValue("ru", NameElement{
+	Expect(profile.Names).To(HaveKeyWithValue("ru", profileapi.NameElement{
 		FirstName:  ptr("Ф"),
 		LastName:   ptr("М"),
 		MiddleName: ptr("Н"),
 	}))
-	Expect(profile.Birth).To(Equal(&EventElement{
-		Date: &DateElement{
+	Expect(profile.Birth).To(Equal(&profileapi.EventElement{
+		Date: &profileapi.DateElement{
 			Day:   ptr[int32](19),
 			Month: ptr[int32](8),
 			Year:  ptr[int32](1922),
 		},
-		Location: &LocationElement{
+		Location: &profileapi.LocationElement{
 			Country:   ptr("РСФСР"),
 			County:    ptr("Спасский уезд, Кирилловская волость"),
 			PlaceName: ptr("село Кирилово"),
@@ -161,8 +163,8 @@ func TestGetProfile2(t *testing.T) {
 		},
 		Name: "Birth of F M",
 	}))
-	Expect(profile.Death).To(Equal(&EventElement{
-		Date: &DateElement{
+	Expect(profile.Death).To(Equal(&profileapi.EventElement{
+		Date: &profileapi.DateElement{
 			Day:   ptr[int32](25),
 			Month: ptr[int32](9),
 			Year:  ptr[int32](1993),
@@ -175,9 +177,9 @@ func TestGetProfile2(t *testing.T) {
 func TestUpdateProfile1(t *testing.T) {
 	t.Skip()
 	RegisterTestingT(t)
-	profileRequest := ProfileRequest{
+	profileRequest := profileapi.Request{
 		Gender: ptr("male"),
-		Names: map[string]NameElement{
+		Names: map[string]profileapi.NameElement{
 			"en-US": {
 				FirstName: ptr("1TestFirstName"),
 				LastName:  ptr("1TestLastName"),
@@ -188,21 +190,21 @@ func TestUpdateProfile1(t *testing.T) {
 				MiddleName: ptr("Н"),
 			},
 		},
-		Birth: &EventElement{
-			Date: &DateElement{
+		Birth: &profileapi.EventElement{
+			Date: &profileapi.DateElement{
 				Day:   ptr[int32](19),
 				Month: ptr[int32](8),
 				Year:  ptr[int32](1922),
 			},
-			Location: &LocationElement{
+			Location: &profileapi.LocationElement{
 				Country:   ptr("РСФСР"),
 				County:    ptr("Спасский уезд, Кирилловская волость"),
 				PlaceName: ptr("село Кирилово"),
 				State:     ptr("Тамбовская губерния"),
 			},
 		},
-		Death: &EventElement{
-			Date: &DateElement{
+		Death: &profileapi.EventElement{
+			Date: &profileapi.DateElement{
 				Day:   ptr[int32](25),
 				Month: ptr[int32](9),
 				Year:  ptr[int32](1993),
@@ -225,22 +227,22 @@ func TestUpdateProfile1(t *testing.T) {
 	Expect(updatedProfile.FirstName).To(BeEquivalentTo("2TestFirstName"))
 	Expect(updatedProfile.LastName).To(BeEquivalentTo("1TestLastName"))
 	Expect(updatedProfile.Gender).To(BeEquivalentTo("male"))
-	Expect(updatedProfile.Names).To(HaveKeyWithValue("en-US", NameElement{
+	Expect(updatedProfile.Names).To(HaveKeyWithValue("en-US", profileapi.NameElement{
 		FirstName: ptr("2TestFirstName"),
 		LastName:  ptr("1TestLastName"),
 	}))
-	Expect(updatedProfile.Names).To(HaveKeyWithValue("ru", NameElement{
+	Expect(updatedProfile.Names).To(HaveKeyWithValue("ru", profileapi.NameElement{
 		FirstName:  ptr("Ф"),
 		LastName:   ptr("М"),
 		MiddleName: ptr("Н"),
 	}))
-	Expect(updatedProfile.Birth).To(Equal(&EventElement{
-		Date: &DateElement{
+	Expect(updatedProfile.Birth).To(Equal(&profileapi.EventElement{
+		Date: &profileapi.DateElement{
 			Day:   ptr[int32](19),
 			Month: ptr[int32](8),
 			Year:  ptr[int32](1922),
 		},
-		Location: &LocationElement{
+		Location: &profileapi.LocationElement{
 			Country:   ptr("РСФСР"),
 			County:    ptr("Спасский уезд, Кирилловская волость"),
 			PlaceName: ptr("село Кирилово"),
@@ -248,8 +250,8 @@ func TestUpdateProfile1(t *testing.T) {
 		},
 		Name: "Birth of 2TestFirstName 1TestLastName",
 	}))
-	Expect(updatedProfile.Death).To(Equal(&EventElement{
-		Date: &DateElement{
+	Expect(updatedProfile.Death).To(Equal(&profileapi.EventElement{
+		Date: &profileapi.DateElement{
 			Day:   ptr[int32](25),
 			Month: ptr[int32](9),
 			Year:  ptr[int32](1993),